Restaurant API

+++ src/main/java/com/swm/sprint1/service/UserService.java
@@ -1,14 +1,13 @@
 package com.swm.sprint1.service;
 
-import com.swm.sprint1.domain.AuthProvider;
-import com.swm.sprint1.domain.User;
-import com.swm.sprint1.domain.UserCategory;
+import com.swm.sprint1.domain.*;
 import com.swm.sprint1.exception.ResourceNotFoundException;
 import com.swm.sprint1.payload.request.UpdateUserCategoryRequest;
 import com.swm.sprint1.payload.request.SignUpRequest;
+import com.swm.sprint1.payload.response.ApiResponse;
 import com.swm.sprint1.repository.CategoryRepository;
-import com.swm.sprint1.repository.UserCategoryRepository;
-import com.swm.sprint1.repository.UserRepository;
+import com.swm.sprint1.repository.user.UserCategoryRepository;
+import com.swm.sprint1.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -43,15 +42,16 @@ public class UserService {
     }
 
     @Transactional
-    public void updateUserCategories(Long id, UpdateUserCategoryRequest updateUserCategoryRequest) {
-        User user = userRepository.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException("User", "id", id));
-
+    public ApiResponse updateUserCategories(User user , UpdateUserCategoryRequest request) {
+        List<Category> categoryByCategoryName = categoryRepository.findCategoryByCategoryName(request.getCategories());
+        if(categoryByCategoryName.isEmpty()){
+            return new ApiResponse(false,"카테고리를 찾을 수 없습니다");
+        }
         userCategoryRepository.deleteUserCategory(user.getId());
-        Set<UserCategory> userCategories = updateUserCategoryRequest.getCategories().stream()
-                .map(name -> new UserCategory(categoryRepository.findByName(name))).collect(Collectors.toSet());
+        Set<UserCategory> userCategories = categoryByCategoryName.stream().map(category -> new UserCategory(user, category)).collect(Collectors.toSet());
         user.changeUserCategory(userCategories);
         userCategories.forEach(userCategory -> userCategoryRepository.save(userCategory));
+        return new ApiResponse(true,"카테고리 수정 완료");
     }
 
     public List<String> getUserCategoryName(Long id) {
